requirements.txt
# Backend & API
fastapi
uvicorn
python-dotenv
pydantic>=2.0.0
python-multipart>=0.0.7

# AI, LLM & RAG
openai
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-core
chromadb
tiktoken

# Data Processing & Visualization
pandas
numpy
plotly
fpdf2
pypdf>=3.0.0

# Frontend
streamlit
requests

# Testing
pytest
httpx


# # ---------------- BACKEND ----------------
# fastapi==0.115.0
# uvicorn==0.30.6
# python-dotenv==1.0.1
# pydantic==2.9.2
# python-multipart==0.0.9

# # ---------------- AI / LLM ----------------
# openai>=1.104.2,<2.0.0
# langchain==0.3.28
# langchain-openai==0.3.35
# langchain-community==0.3.31
# langchain-core==0.3.83
# langchain-chroma==0.2.6
# chromadb==1.5.5
# tiktoken==0.12.0

# # ---------------- DATA ----------------
# pandas==2.2.2
# numpy==1.26.4
# plotly==6.0.0
# fpdf2==2.8.7
# pypdf==6.0.0

# # ---------------- FRONTEND ----------------
# streamlit==1.39.0
# requests==2.32.3

# # ---------------- TESTING ----------------
# pytest==8.3.3
# httpx==0.27.2

# # ---------------- FIX (IMPORTANT) ----------------
# # Force compatible OpenTelemetry (prevents your error)
# opentelemetry-api==1.37.0
# opentelemetry-sdk==1.37.0
# opentelemetry-proto==1.37.0
# opentelemetry-exporter-otlp-proto-common==1.37.0
# opentelemetry-exporter-otlp-proto-http==1.37.0
# opentelemetry-semantic-conventions==0.58b0